+++ test/rand_sub.c
@@ -33,7 +33,7 @@ static void OnPubMatch(DPS_Subscription* sub, const DPS_Publication* pub, uint8_
         }
         DPS_PRINT("\n");
         if (data) {
-            DPS_PRINT("%.*s\n", len, data);
+            DPS_PRINT("%.*s\n", (int)len, data);
         }
     }
 }
